move merkle and signing/legacy into mam package
