+++ mam/v1/mam_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/iotaledger/iota.go/converter"
 	. "github.com/iotaledger/iota.go/curl"
 	. "github.com/iotaledger/iota.go/mam/v1"
-	. "github.com/iotaledger/iota.go/merkle"
+	. "github.com/iotaledger/iota.go/mam/v1/merkle"
 	"github.com/iotaledger/iota.go/trinary"
 )
