+++ mam/v1/signing/signing.go
