+++ mam/v1/signing/signing_test.go
@@ -3,7 +3,7 @@ package signing_test
 import (
 	. "github.com/iotaledger/iota.go/consts"
 	. "github.com/iotaledger/iota.go/curl"
-	. "github.com/iotaledger/iota.go/signing/legacy"
+	. "github.com/iotaledger/iota.go/mam/v1/signing"
 	. "github.com/iotaledger/iota.go/trinary"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
