+++ signing/signing_milestone_test.go
@@ -3,7 +3,7 @@ package signing_test
 import (
 	"github.com/iotaledger/iota.go/consts"
 	. "github.com/iotaledger/iota.go/curl"
-	"github.com/iotaledger/iota.go/merkle"
+	"github.com/iotaledger/iota.go/mam/v1/merkle"
 	. "github.com/iotaledger/iota.go/signing"
 	. "github.com/iotaledger/iota.go/transaction"
 	. "github.com/iotaledger/iota.go/trinary"
